+++ examples/next-openai/app/api/use-chat-streamdata/route.ts
@@ -1,35 +1,38 @@
 import { openai } from '@ai-sdk/openai';
-import { generateId, StreamData, streamText } from 'ai';
-
-// Allow streaming responses up to 30 seconds
-export const maxDuration = 30;
+import { generateId, createDataStreamResponse, streamText } from 'ai';
 
 export async function POST(req: Request) {
   const { messages } = await req.json();
 
-  // use stream data
-  const data = new StreamData();
-  data.append('initialized call');
+  // immediately start streaming (solves RAG issues with status, etc.)
+  return createDataStreamResponse({
+    execute: dataStream => {
+      dataStream.writeData('initialized call');
 
-  const result = streamText({
-    model: openai('gpt-4o'),
-    messages,
-    onChunk() {
-      data.appendMessageAnnotation({ chunk: '123' });
-    },
-    onFinish() {
-      // message annotation:
-      data.appendMessageAnnotation({
-        id: generateId(), // e.g. id from saved DB record
-        other: 'information',
+      const result = streamText({
+        model: openai('gpt-4o'),
+        messages,
+        onChunk() {
+          dataStream.writeMessageAnnotation({ chunk: '123' });
+        },
+        onFinish() {
+          // message annotation:
+          dataStream.writeMessageAnnotation({
+            id: generateId(), // e.g. id from saved DB record
+            other: 'information',
+          });
+
+          // call annotation:
+          dataStream.writeData('call completed');
+        },
       });
 
-      // call annotation:
-      data.append('call completed');
-
-      data.close();
+      result.mergeIntoDataStream(dataStream);
+    },
+    onError: error => {
+      // Error messages are masked by default for security reasons.
+      // If you want to expose the error message to the client, you can do so here:
+      return error instanceof Error ? error.message : String(error);
     },
   });
-
-  return result.toDataStreamResponse({ data });
 }