+++ packages/ai/streams/langchain-adapter.ts
@@ -1,10 +1,12 @@
+import { formatDataStreamPart } from '@ai-sdk/ui-utils';
+import { DataStreamWriter } from '../core/data-stream/data-stream-writer';
 import { mergeStreams } from '../core/util/merge-streams';
 import { prepareResponseHeaders } from '../core/util/prepare-response-headers';
 import {
   createCallbacksTransformer,
   StreamCallbacks,
 } from './stream-callbacks';
-import { createStreamDataTransformer, StreamData } from './stream-data';
+import { StreamData } from './stream-data';
 
 type LangChainImageDetail = 'auto' | 'low' | 'high';
 
@@ -45,14 +47,7 @@ type LangChainStreamEvent = {
   data: any;
 };
 
-/**
-Converts LangChain output streams to an AI SDK Data Stream.
-
-The following streams are supported:
-- `LangChainAIMessageChunk` streams (LangChain `model.stream` output)
-- `string` streams (LangChain `StringOutputParser` output)
- */
-export function toDataStream(
+function toDataStreamInternal(
   stream:
     | ReadableStream<LangChainStreamEvent>
     | ReadableStream<LangChainAIMessageChunk>
@@ -89,7 +84,33 @@ export function toDataStream(
       }),
     )
     .pipeThrough(createCallbacksTransformer(callbacks))
-    .pipeThrough(createStreamDataTransformer());
+    .pipeThrough(new TextDecoderStream())
+    .pipeThrough(
+      new TransformStream({
+        transform: async (chunk, controller) => {
+          controller.enqueue(formatDataStreamPart('text', chunk));
+        },
+      }),
+    );
+}
+
+/**
+Converts LangChain output streams to an AI SDK Data Stream.
+
+The following streams are supported:
+- `LangChainAIMessageChunk` streams (LangChain `model.stream` output)
+- `string` streams (LangChain `StringOutputParser` output)
+ */
+export function toDataStream(
+  stream:
+    | ReadableStream<LangChainStreamEvent>
+    | ReadableStream<LangChainAIMessageChunk>
+    | ReadableStream<string>,
+  callbacks?: StreamCallbacks,
+) {
+  return toDataStreamInternal(stream, callbacks).pipeThrough(
+    new TextEncoderStream(),
+  );
 }
 
 export function toDataStreamResponse(
@@ -103,7 +124,10 @@ export function toDataStreamResponse(
     callbacks?: StreamCallbacks;
   },
 ) {
-  const dataStream = toDataStream(stream, options?.callbacks);
+  const dataStream = toDataStreamInternal(
+    stream,
+    options?.callbacks,
+  ).pipeThrough(new TextEncoderStream());
   const data = options?.data;
   const init = options?.init;
 
@@ -119,6 +143,16 @@ export function toDataStreamResponse(
       dataStreamVersion: 'v1',
     }),
   });
+}
+
+export function mergeIntoDataStream(
+  stream:
+    | ReadableStream<LangChainStreamEvent>
+    | ReadableStream<LangChainAIMessageChunk>
+    | ReadableStream<string>,
+  options: { dataStream: DataStreamWriter; callbacks?: StreamCallbacks },
+) {
+  options.dataStream.merge(toDataStreamInternal(stream, options.callbacks));
 }
 
 function forwardAIMessageChunk(